@@ -25,9 +25,7 @@
     "babel-preset-env": "^1.6.1",
     "file-loader": "^0.10.1",
     "redux-devtools": "^3.3.2",
-    "semantic-ui": "^2.4.0",
-    "webpack": "^3.0.0",
-    "webpack-dev-server": "^2.4.1"
+    "webpack": "^3.0.0"
   },
   "dependencies": {
     "babel-loader": "^6.4.0",
@@ -36,36 +34,37 @@
     "babel-preset-es2015": "^6.22.0",
     "babel-preset-react": "^6.23.0",
     "babel-preset-stage-1": "^6.24.1",
-    "babel-preset-stage-2": "^6.24.1",
-    "bootstrap": "^4.0.0-beta",
+    "babel-preset-stage-2": "^6.22.0",
+    "better-console": "^1.0.1",
     "classnames": "^2.2.5",
     "clean-webpack-plugin": "^0.1.19",
     "css-loader": "^0.28.7",
+    "gulp": "^3.9.1",
+    "gulp-autoprefixer": "^6.0.0",
+    "gulp-help": "^1.6.1",
     "history": "1.13.x",
     "html-webpack-plugin": "^3.0.7",
     "lodash": "^4.17.5",
     "natives": "^1.1.4",
-    "react": "^15.6.2",
+    "react": "^16.5.2",
     "react-addons-perf": "^15.4.2",
-    "react-bootstrap": "^0.30.8",
-    "react-dom": "^15.6.2",
+    "react-dom": "^16.5.2",
     "react-notification-system": "^0.2.15",
     "react-notification-system-redux": "^1.1.6",
     "react-paginate": "^4.4.2",
-    "react-portal": "^4.1.3",
     "react-redux": "^5.0.3",
     "react-router": "latest",
     "react-router-dom": "^4.0.0-beta.7",
     "react-select": "^1.0.0-rc.5",
     "react-transition-group": "^1.2.1",
-    "reactstrap": "^5.0.0-alpha.3",
     "redux": "^3.6.0",
     "redux-notifications": "^4.0.1",
+    "require-dot-file": "^0.4.0",
     "reselect": "^3.0.1",
-    "semantic-ui-react": "^0.75.0",
+    "semantic-ui": "^2.4.0",
+    "semantic-ui-react": "^0.82.5",
     "socket.io-client": "^1.7.3",
     "style-loader": "^0.19.0",
-    "url-loader": "^1.0.1",
-    "yarn": "^1.5.1"
+    "url-loader": "^1.0.1"
   }
 }
