@@ -1,5 +1,6 @@
 import _ from 'lodash'
 import React from 'react'
+import PropTypes from 'prop-types';
 import { Dimmer, Loader, Segment } from 'semantic-ui-react'
 
 import HostsList from './HostsList.jsx'
@@ -90,7 +91,7 @@ class HostsListScopesUpdater extends React.Component {
 }
 
 HostsListScopesUpdater.contextTypes = {
-    store: React.PropTypes.object
+    store: PropTypes.object
 }
 
 export default HostsListScopesUpdater;
